+++ packages/target-tauri/src-tauri/src/cli.rs
@@ -8,6 +8,15 @@ use crate::run_config::RunConfig;
 #[derive(Parser)]
 #[command(about, long_about = None)]
 pub struct Cli {
+    #[cfg(target_os = "macos")]
+    /// Start deltachat in minimized mode
+    #[arg(short, long)]
+    minimized: bool,
+    #[cfg(not(target_os = "macos"))]
+    /// Start deltachat in minimized mode with tray icon
+    /// (tray icon will be activated for this session regardless whether it's disabled)
+    #[arg(short, long)]
+    minimized: bool,
     /// opens devtools and activates --log-debug & --log-to-console
     #[arg(long)]
     dev_mode: bool,
@@ -17,6 +26,10 @@ pub struct Cli {
     /// Output the log to stdout / Browser dev console
     #[arg(long)]
     log_to_console: bool,
+    /// enable auto-reload for translations.
+    /// You can use it in combination with the env var `DELTACHAT_LOCALE_DIR`.
+    #[arg(long)]
+    watch_translations: bool,
     /// Print version
     #[arg(short = 'V', long)]
     version: bool,
@@ -55,5 +68,10 @@ Webview: {:?}",
         log_to_console: cli.dev_mode || cli.log_to_console,
         devtools: cli.dev_mode,
         dev_mode: cli.dev_mode,
+        translation_watch: cli.dev_mode || cli.watch_translations,
+        minimized_window: cli.minimized,
+        // Tray icon is not forced on macOS because the app icon is still in the Dock,
+        // even when Delta Chat has no visible window
+        forced_tray_icon: cli.minimized && !cfg!(target_os = "macos"),
     }
 }