+++ packages/target-tauri/src-tauri/src/state/menu_manager.rs
@@ -1,6 +1,5 @@
 use std::{collections::HashMap, sync::Arc};
 
-use anyhow::Context;
 use log::error;
 use tauri::{
     async_runtime::spawn, AppHandle, Manager, Runtime, WebviewWindow, Window, WindowEvent, Wry,
@@ -9,6 +8,8 @@ use tokio::sync::RwLock;
 
 #[cfg(desktop)]
 use tauri::menu::Menu;
+
+use crate::TrayManager;
 
 #[cfg(desktop)]
 type GenerateMenuFn = Box<dyn Fn(&AppHandle) -> anyhow::Result<Menu<Wry>> + Send + Sync>;
@@ -82,6 +83,8 @@ impl MenuManager {
         app: &AppHandle,
         id: &str,
     ) -> anyhow::Result<Menu<Wry>> {
+        use anyhow::Context;
+
         let inner = self.inner.read().await;
         let menu_generator = inner
             .get(id)
@@ -154,6 +157,9 @@ impl MenuManager {
             if let Err(err) = cloned_self.inner_update_all(&app).await {
                 error!("error while updating all menus: {err:?}");
             }
+            if let Err(err) = app.state::<TrayManager>().update_menu(&app).await {
+                error!("error while updating tray menu: {err:?}");
+            }
         });
     }
 
@@ -177,6 +183,8 @@ impl MenuManager {
         // but on macOS only set the global menu for the focused window
         #[cfg(target_os = "macos")]
         {
+            use anyhow::Context;
+
             let win = app
                 .get_focused_window()
                 .or(app.get_window("main"))