+++ packages/target-tauri/src-tauri/src/state/mod.rs
@@ -4,4 +4,6 @@ pub(crate) mod html_email_instances;
 pub(crate) mod main_window_channels;
 pub(crate) mod menu_manager;
 pub(crate) mod translations;
+#[cfg(desktop)]
+pub(crate) mod tray_manager;
 pub(crate) mod webxdc_instances;