+++ packages/target-tauri/src-tauri/src/util/mod.rs
@@ -2,4 +2,7 @@ pub mod csp;
 pub mod sanitization;
 mod truncate_text;
 
+#[cfg(desktop)]
+pub(crate) mod fs_watcher;
+
 pub(crate) use truncate_text::truncate_text;